# performance: Implement sync.Pool for cow.child()