--- a/ledger/internal/eval.go
+++ b/ledger/internal/eval.go
@@ -934,7 +934,7 @@ func (eval *BlockEvaluator) transactionGroup(txgroup []transactions.SignedTxnWit
 	var groupTxBytes int
 
 	cow := eval.state.child(len(txgroup))
-	defer childPool.Put(cow)
+	defer cow.recycle()
 
 	evalParams := logic.NewEvalParams(txgroup, &eval.proto, &eval.specials)