--- a/ledger/ledgercore/statedelta.go
+++ b/ledger/ledgercore/statedelta.go
@@ -87,6 +87,7 @@ type KvValueDelta struct {
 // StateDelta describes the delta between a given round to the previous round
 // If adding a new field not explicitly allocated by PopulateStateDelta, make sure to reset
 // it in .ReuseStateDelta to avoid dirty memory errors.
+// If adding fields make sure to add them to the .Reset() method to avoid dirty state
 type StateDelta struct {
 	// modified new accounts
 	Accts AccountDeltas
@@ -175,6 +176,7 @@ type AssetResourceRecord struct {
 // do that, each of the arrays here is constructed as a pair of (slice, map).
 // The map would point the address/address+creatable id onto the index of the
 // element within the slice.
+// If adding fields make sure to add them to the .reset() method to avoid dirty state
 type AccountDeltas struct {
 	// Actual data. If an account is deleted, `Accts` contains the BalanceRecord
 	// with an empty `AccountData` and a populated `Addr`.
@@ -204,11 +206,13 @@ func MakeStateDelta(hdr *bookkeeping.BlockHeader, prevTimestamp int64, hint int,
 
 // PopulateStateDelta populates an existing StateDelta struct.
 // Used as a helper for MakeStateDelta as well as for re-using already allocated structs from sync.Pool
-// Make sure to .Reset the struct before re-use
 func (sd *StateDelta) PopulateStateDelta(hdr *bookkeeping.BlockHeader, prevTimestamp int64, hint int, stateProofNext basics.Round) {
-	sd.Accts = MakeAccountDeltas(hint)
-	sd.Txids = make(map[transactions.Txid]IncludedTransactions, hint)
-	// asset or application creation are considered as rare events so do not pre-allocate space for them
+	if sd.Txids == nil {
+		sd.Txids = make(map[transactions.Txid]IncludedTransactions, hint)
+	}
+	if sd.Accts.isEmpty() {
+		sd.Accts = MakeAccountDeltas(hint)
+	}
 	sd.Hdr = hdr
 	sd.StateProofNext = stateProofNext
 	sd.PrevTimestamp = prevTimestamp
@@ -216,6 +220,7 @@ func (sd *StateDelta) PopulateStateDelta(hdr *bookkeeping.BlockHeader, prevTimes
 }
 
 // MakeAccountDeltas creates account delta
+// if adding new fields make sure to add them to the .reset() and .isEmpty() methods
 func MakeAccountDeltas(hint int) AccountDeltas {
 	return AccountDeltas{
 		Accts:      make([]BalanceRecord, 0, hint*2),
@@ -223,14 +228,52 @@ func MakeAccountDeltas(hint int) AccountDeltas {
 	}
 }
 
-// ReuseStateDelta resets fields that aren't directly set by PopulateStateDelta
-func (sd *StateDelta) ReuseStateDelta(hdr *bookkeeping.BlockHeader, prevTimestamp int64, hint int, stateProofNext basics.Round) {
-	sd.PopulateStateDelta(hdr, prevTimestamp, hint, stateProofNext)
-	// Reset values that aren't explicitly set by PopulateStateDelta
-	sd.Txleases = nil
-	sd.Creatables = nil
-	sd.KvMods = nil
+// Reset resets the StateDelta for re-use with sync.Pool
+func (sd *StateDelta) Reset() {
+	sd.Accts.reset()
+	for txid := range sd.Txids {
+		delete(sd.Txids, txid)
+	}
+	for txLease := range sd.Txleases {
+		delete(sd.Txleases, txLease)
+	}
+	for creatableIndex := range sd.Creatables {
+		delete(sd.Creatables, creatableIndex)
+	}
+	for key := range sd.KvMods {
+		delete(sd.KvMods, key)
+	}
 	sd.Totals = AccountTotals{}
+
+	// these fields are going to be populated on next use but resetting them anyway for safety.
+	sd.Hdr = nil
+	sd.StateProofNext = basics.Round(0)
+	sd.PrevTimestamp = 0
+	sd.initialTransactionsCount = 0
+}
+
+// reset clears out allocated slices from AccountDeltas struct for reuse with sync.Pool
+func (ad *AccountDeltas) reset() {
+	// reset the slices
+	ad.Accts = ad.Accts[:0]
+	ad.AppResources = ad.AppResources[:0]
+	ad.AssetResources = ad.AssetResources[:0]
+
+	// reset the maps
+	for address := range ad.acctsCache {
+		delete(ad.acctsCache, address)
+	}
+	for aApp := range ad.appResourcesCache {
+		delete(ad.appResourcesCache, aApp)
+	}
+	for aAsset := range ad.assetResourcesCache {
+		delete(ad.assetResourcesCache, aAsset)
+	}
+}
+
+// isEmpty returns true if any of the fields allocated by MakeAccountDeltas is nil
+func (ad *AccountDeltas) isEmpty() bool {
+	return ad.Accts == nil || ad.acctsCache == nil
 }
 
 // GetData lookups AccountData by address