## Summary

Use sync.pool for `roundCowState`s created through `child()` method. 

## Test Plan

- [x] Passes existing tests
- [x] Add new test to protect against "dirty" state being returned from the pool and propagated

## Performance improvements
scenario1s alloc_objects:
recent tx_pointer run (not touching child() codepath): **11.6%    183354536**
this change: **3.8%    76110520**
~ 60% decrease in number of allocs through `child()`

BenchmarkCowChild results (on master commented out call to child.recycle() since it's not implemented there)
```
old - BenchmarkCowChild-10    	   95524	     12138 ns/op	  273075 B/op	       5 allocs/op
new - BenchmarkCowChild-10    	15723090	        70.02 ns/op	       0 B/op	       0 allocs/op

```

Updated benchmarks with the max possible hint of 16:
```
name         old time/op    new time/op    delta
CowChild-10    1.15µs ± 7%    0.07µs ± 1%   -93.89%  (p=0.000 n=20+19)

name         old alloc/op   new alloc/op   delta
CowChild-10    14.7kB ± 0%     0.0kB       -100.00%  (p=0.000 n=20+20)

name         old allocs/op  new allocs/op  delta
CowChild-10      3.00 ± 0%      0.00       -100.00%  (p=0.000 n=20+20)"
```